pallet_maker_backend/requirements.txt
fastapi
uvicorn
asyncua
python-dotenv